requirements_xserver.txt
Flask==3.1.1
flask-cors==6.0.0
SQLAlchemy==2.0.25
PyMySQL==1.1.0
python-dotenv==1.1.0
cryptography==41.0.7